chore: adds generic add_prop func to package-level logic for reuse
